chore(TracklistsModule): reorganize file structure
