+++ src/app/player/components/player-controls.ts
@@ -1,6 +1,6 @@
 import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output, ViewEncapsulation } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
-import { Track, TracklistCursor } from 'src/tracklists';
+import { Track, TracklistCursor } from 'src/app/tracklists';
 import { PlayerState } from '../state/player-state';
