+++ src/app/player/state/selectors.ts
@@ -7,7 +7,7 @@ import 'rxjs/add/operator/withLatestFrom';
 
 import { AppState } from 'src/app';
 import { Selector } from 'src/app/core';
-import { getTracklistCursor, getTracklists, getTracks, Track, Tracklist, TracklistCursor } from 'src/tracklists';
+import { getTracklistCursor, getTracklists, getTracks, Track, Tracklist, TracklistCursor } from 'src/app/tracklists';
 import { PlayerState } from './player-state';
 import { TimesState } from './times-state';
