+++ src/app/tracklists/components/waveform-timeline.scss
