+++ src/app/tracklists/components/waveform.scss
