+++ src/app/tracklists/tracklists-module.ts
@@ -2,26 +2,20 @@ import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { EffectsModule } from '@ngrx/effects';
 
-import { SharedModule } from 'src/app/shared';
-
+// components
 import { TrackCardComponent } from './components/track-card';
 import { TracklistComponent } from './components/tracklist';
 import { TracklistItemsComponent } from './components/tracklist-items';
 import { WaveformComponent } from './components/waveform';
 import { WaveformTimelineComponent } from './components/waveform-timeline';
 
+// modules
+import { SharedModule } from '../shared';
+
+// services
 import { TracklistActions } from './tracklist-actions';
 import { TracklistEffects } from './tracklist-effects';
 import { TracklistService } from './tracklist-service';
-
-
-export { TracklistActions, TracklistService };
-export { createTrack, Track, TrackData, TrackRecord } from './models/track';
-export { Tracklist, TracklistRecord } from './models/tracklist';
-export { getCurrentTracklist, getTracks, getTracklists } from './reducers/selectors';
-export { TracklistsState, tracklistsReducer } from './reducers/tracklists-reducer';
-export { TracksState, tracksReducer } from './reducers/tracks-reducer';
-export { getTracklistCursor, TracklistCursor } from './tracklist-cursor';
 
 
 @NgModule({
