+++ src/users/reducers/selectors.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from '@angular/core/testing';
 import { Store, StoreModule } from '@ngrx/store';
-import { TracklistActions } from 'src/tracklists';
+import { TracklistActions } from 'src/app/tracklists';
 import { testUtils } from 'src/utils/test';
 import { createUser } from '../models/user';
 import { UserActions } from '../user-actions';
